+++ src/Controller/PageController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\Book;
 use App\Entity\Page;
 use App\Form\PageType;
 use App\Repository\PageRepository;
@@ -10,29 +11,21 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
- * @Route("/page")
+ * @Route("book/{slug}/page")
  */
 class PageController extends AbstractController
 {
     /**
-     * @Route("/", name="page_index", methods={"GET"})
-     */
-    public function index(PageRepository $pageRepository): Response
-    {
-        return $this->render('page/index.html.twig', [
-            'pages' => $pageRepository->findAll(),
-        ]);
-    }
-
-    /**
      * @Route("/new", name="page_new", methods={"GET","POST"})
+     * @param Book $book
      * @param Request $request
      * @param Slugify $slugify
      * @return Response
      */
-    public function new(Request $request, Slugify $slugify): Response
+    public function new(Book $book, Request $request, Slugify $slugify): Response
     {
         $page = new Page();
         $form = $this->createForm(PageType::class, $page);
@@ -40,47 +33,53 @@ class PageController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $slug = $slugify->generate($page->getTitle());
-            $page->setSlug($slug);
+            $page->setSlug($slug)
+                ->setBook($book)
+                ->setPageNumber(count($book->getPages()) + 1);
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($page);
             $entityManager->flush();
 
-            return $this->redirectToRoute('page_index');
+            return $this->redirectToRoute('book_show', [
+                'slug' => $book->getSlug(),
+            ]);
         }
 
         return $this->render('page/new.html.twig', [
             'page' => $page,
             'form' => $form->createView(),
+            'book' => $book,
         ]);
     }
 
     /**
-     * @Route("/{slug}", name="page_show", methods={"GET"})
+     * @Route("/{id}/edit", name="page_edit", methods={"GET","POST"})
+     * @param Book $book
+     * @param Request $request
+     * @param Page $page
+     * @return Response
      */
-    public function show(Page $page): Response
+    public function edit(Book $book, Request $request, Page $page): Response
     {
-        return $this->render('page/show.html.twig', [
-            'page' => $page,
-        ]);
-    }
+        if (!($this->getUser() == $page->getBook()->getAuthor())) {
+            throw new AccessDeniedException('Only the author can edit this page!');
+        }
 
-    /**
-     * @Route("/{slug}/edit", name="page_edit", methods={"GET","POST"})
-     */
-    public function edit(Request $request, Page $page): Response
-    {
         $form = $this->createForm(PageType::class, $page);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('page_index');
+            return $this->redirectToRoute('book_show', [
+                'slug' => $book->getSlug(),
+            ]);
         }
 
         return $this->render('page/edit.html.twig', [
             'page' => $page,
             'form' => $form->createView(),
+            'book' => $book,
         ]);
     }
 
@@ -89,12 +88,16 @@ class PageController extends AbstractController
      */
     public function delete(Request $request, Page $page): Response
     {
+        if (!($this->getUser() == $page->getBook()->getAuthor())) {
+            throw new AccessDeniedException('Only the author can delete this page!');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$page->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($page);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('page_index');
+        return $this->redirectToRoute('book_index');
     }
 }
